Include suspended network when getAllNetworkStateSnapshots
Suspended network should be considered as temporary shortage of
connectivity of a connected network. Thus, it should not be
excluded from network state snapshots and causes data usage to
stop accounting or iptables rules to be removed on the interface
of the suspended network.

This change also address the naming confusion of default networks
parameter of expectNotifyNetworkStatus.

Test: atest ConnectivityServiceTest#testGetAllNetworkStateSnapshots
Bug: 232479679
Bug: 196079981
(cherry picked from commit ddecb662463020661858f3219d405dce1dd116eb)
Change-Id: Ie2d9c2672e8bfd57343e82242e387b658a299b97